move main and grid_search to bin/
